.claude/worktrees/agitated-carson-a7fd6e/backend/requirements.txt
Django==4.2.10
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
django-oauth-toolkit==2.3.0
psycopg2-binary==2.9.9
redis==5.0.1
python-dotenv==1.0.0
pyotp==2.9.0
cryptography==41.0.7
stripe==7.5.0
boto3==1.34.23
django-storages==1.14.2
requests==2.31.0
django-ratelimit==4.1.0
django-cors-headers==4.3.1
gunicorn==21.2.0
qrcode==7.4.2
Pillow ==10.1.0
supabase==2.28.0
dj-database-url==3.1.2
django-phonenumber-field==7.3.0
phonenumbers==8.13.55